refactor(mall-distribution): 重构分销领域模型结构
