+++ mall-distribution-application/src/main/java/com/mall/distribution/application/service/DistributorJoinService.java
@@ -4,12 +4,12 @@ import com.mall.distribution.application.command.CreateDistributorJoinCommand;
 import com.mall.distribution.application.port.DistributorJoinUseCase;
 import com.mall.distribution.application.query.DistributorJoinQuery;
 import com.mall.distribution.common.exception.BusinessException;
-import com.mall.distribution.common.exception.Resp;
-import com.mall.distribution.domain.distributor.DistributorJoin;
-import com.mall.distribution.domain.distributor.DistributorJoinRepository;
+
+import com.mall.distribution.domain.model.join.DistributorJoin;
+import com.mall.distribution.domain.repository.DistributorJoinRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+
 
 import java.util.List;
 
@@ -19,9 +19,9 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class DistributorJoinService implements DistributorJoinUseCase {
-    
+
     private final DistributorJoinRepository distributorJoinRepository;
-    
+
     @Override
     @Transactional
     public Long createDistributorJoin(CreateDistributorJoinCommand command) {
@@ -29,7 +29,7 @@ public class DistributorJoinService implements DistributorJoinUseCase {
         if (distributorJoinRepository.existsPendingByUserId(command.getUserId())) {
             throw new BusinessException(Resp.DISTRIBUTOR_JOIN_PENDING_EXISTS);
         }
-        
+
         // 创建申请
         DistributorJoin distributorJoin = DistributorJoin.create(
             command.getUserId(),
@@ -44,24 +44,24 @@ public class DistributorJoinService implements DistributorJoinUseCase {
             command.getReason(),
             command.getCreatedBy()
         );
-        
+
         // 保存申请
         distributorJoinRepository.save(distributorJoin);
-        
+
         return distributorJoin.getJoinId();
     }
-    
+
     @Override
     public DistributorJoin findById(Long joinId) {
         return distributorJoinRepository.findById(joinId)
             .orElseThrow(() -> new BusinessException(Resp.DISTRIBUTOR_JOIN_NOT_FOUND));
     }
-    
+
     @Override
     public List<DistributorJoin> findDistributorJoins(DistributorJoinQuery query) {
         return distributorJoinRepository.findByQuery(query);
     }
-    
+
     @Override
     @Transactional
     public void approveJoin(Long joinId, String auditBy, String auditRemark) {
@@ -69,7 +69,7 @@ public class DistributorJoinService implements DistributorJoinUseCase {
         distributorJoin.approve(auditBy, auditRemark);
         distributorJoinRepository.save(distributorJoin);
     }
-    
+
     @Override
     @Transactional
     public void rejectJoin(Long joinId, String auditBy, String auditRemark) {
@@ -77,7 +77,7 @@ public class DistributorJoinService implements DistributorJoinUseCase {
         distributorJoin.reject(auditBy, auditRemark);
         distributorJoinRepository.save(distributorJoin);
     }
-    
+
     @Override
     public boolean existsPendingByUserId(Long userId) {
         return distributorJoinRepository.existsPendingByUserId(userId);
